+++ plugins/terraform/resource_layer0_load_balancer.go
@@ -118,7 +118,7 @@ func resourceLayer0LoadBalancerCreate(d *schema.ResourceData, meta interface{}) 
 	private := d.Get("private").(bool)
 	ports := expandPorts(d.Get("port").(*schema.Set).List())
 	healthCheck := expandHealthCheck(d.Get("health_check"))
-	idleTimeout := d.Get("idle_timeout").(int64)
+	idleTimeout := d.Get("idle_timeout").(int)
 
 	if healthCheck == nil {
 		healthCheck = &models.HealthCheck{
